For the new compiled version fix yet another bug
The vector was improperly indexed in the Coulomb coefficient
method. Also begin implementing post-HF methods. For the H2
calculation in the Python implementation, the method
that uses one primitive per basis function is no longer used.

+++ gaussian_basis/gaussian_basis/src2/basis_function_array.cpp
@@ -12,7 +12,8 @@ Gaussian3D BasisFunctionArray::primitive(const BasisFunctionData &b, int i) cons
     spatial::Vector position = b.position;
     spatial::UByte4 angular = b.angular;
     return Gaussian3D(
-        exponent, amplitude, angular.x, angular.y, angular.z, position);
+        exponent, amplitude, 
+        (int)angular.x, (int)angular.y, (int)angular.z, position);
 }
 
 BasisFunctionArray::BasisFunctionArray(
